Aihc.Parser

Description

This module provides parsing functions for Haskell source code. The main entry point is parseModule for parsing complete Haskell modules. Additional functions are provided for parsing individual expressions, patterns, and types.

Parsing modules

parseModule :: ParserConfig -> Text -> ([(SourceSpan, Text)], Module)

Parse a complete Haskell module.

Returns any recovered parse errors alongside a (possibly partial) Module. When individual declarations fail to parse, the parser recovers and continues, returning the error and the successfully parsed declarations.

>>> shorthand $ snd $ parseModule defaultConfig "module Main where\nmain = putStrLn \"Hello\""
Module {ModuleHead {"Main"}, [DeclValue (PatternBind (PVar "main") (EApp (EVar "putStrLn") (EString "Hello")))]}

Modules without a header are also supported:

>>> case parseModule defaultConfig "x = 1" of { (_, m) -> moduleName m }
Nothing

Configuration

data ParserConfig

Constructors

ParserConfig

Fields

parserSourceName :: FilePath

parserExtensions :: [Extension]

defaultConfig :: ParserConfig

Default parser configuration.

parserSourceName is set to "<input>"

parserExtensions is empty (no extensions enabled by default)

>>> parserSourceName defaultConfig
"<input>"

>>> parserExtensions defaultConfig
[]

Parse results

data ParseResult a

Constructors

ParseOk a

ParseErr [(SourceSpan, Text)]

formatParseErrors :: FilePath -> Maybe Text -> [(SourceSpan, Text)] -> String

Pretty-print a list of spanned parse errors with source context.

Parsing expressions, patterns, types, and declarations

parseExpr :: ParserConfig -> Text -> ParseResult Expr

Parse a Haskell expression.

>>> shorthand $ parseExpr defaultConfig "1 + 2"
ParseOk (EInfix (EInt 1 TInteger) "+" (EInt 2 TInteger))

>>> shorthand $ parseExpr defaultConfig "\\x -> x + 1"
ParseOk (ELambdaPats [PVar "x"] (EInfix (EVar "x") "+" (EInt 1 TInteger)))

Parse errors are returned as ParseErr:

>>> case parseExpr defaultConfig "1 +" of { ParseErr _ -> "error"; ParseOk _ -> "ok" }
"error"

parseSignatureType :: ParserConfig -> Text -> ParseResult Type

Parse a Haskell signature type.

This is the context used by top-level type signatures. Unlike parseType, it rejects an unparenthesized outer kind signature:

>>> case parseSignatureType defaultConfig "_ :: _" of { ParseErr _ -> "error"; ParseOk _ -> "ok" }
"error"

parseType :: ParserConfig -> Text -> ParseResult Type

Parse a Haskell type in the general declaration RHS context.

>>> shorthand $ parseType defaultConfig "Int -> Bool"
ParseOk (TFun (TCon "Int") (TCon "Bool"))

>>> shorthand $ parseType defaultConfig "Maybe a"
ParseOk (TApp (TCon "Maybe") (TVar "a"))

>>> shorthand $ parseType defaultConfig "_ :: _"
ParseOk (TKindSig (TWildcard) (TWildcard))

parsePattern :: ParserConfig -> Text -> ParseResult Pattern

Parse a Haskell pattern.

>>> shorthand $ parsePattern defaultConfig "(x, y)"
ParseOk (PTuple [PVar "x", PVar "y"])

>>> shorthand $ parsePattern defaultConfig "Just x"
ParseOk (PCon "Just" [PVar "x"])

parseDecl :: ParserConfig -> Text -> ParseResult Decl

Parse a single Haskell declaration.

>>> shorthand $ parseDecl defaultConfig "f x = x + 1"
ParseOk (DeclValue (FunctionBind "f" [Match {MatchHeadPrefix, [PVar "x"], EInfix (EVar "x") "+" (EInt 1 TInteger)}]))