Aihc.Parser.Pretty

Description

This module provides pretty-printing of parsed AST back to valid Haskell source code. It is used for round-trip testing and code generation.

The Pretty instances from Prettyprinter are provided for the main AST types, allowing direct use of pretty from Prettyprinter.

Parenthesization is handled by the Parens module, which inserts EParen, PParen, TParen, and CmdPar nodes at all required positions. The Pretty instances call the paren-insertion pass before formatting, so the formatting code here does not need to worry about operator precedence or context-sensitive parenthesization.

Import this module to bring the Pretty instances into scope. The exported helpers render AST nodes without first inserting additional parentheses, which is useful for tests that need to inspect parenthesization directly.

Provided instances: Pretty for Module, Expr, Pattern, Type.

Documentation

prettyExpr :: Expr -> Doc ann

Pretty-print an expression. The AST is assumed to already have EParen nodes in the correct positions (inserted by addExprParens).

>>> let expr = EApp (EVar "f") (EParen (EInfix (EVar "x") "+" (EVar "y"))) in (renderDoc (prettyExpr expr), show (shorthand expr))
("f\n  (x\n   + y)","EApp (EVar \"f\") (EParen (EInfix (EVar \"x\") \"+\" (EVar \"y\")))")

prettyPattern :: Pattern -> Doc ann

Pretty-print a pattern. The AST is assumed to already have PParen nodes in the correct positions (inserted by addPatternParens).

>>> let pat = PAs "xs" (PParen (PNegLit (LitInt 1 TInteger "1"))) in (renderDoc (prettyPattern pat), show (shorthand pat))
("xs@(-1)","PAs UnqualifiedName {\"xs\"} (PParen (PNegLit (LitInt 1 TInteger)))")

prettyType :: Type -> Doc ann

Orphan instances

Pretty Decl
Pretty instance for Decl - renders to valid Haskell source code.

Instance details

Methods

pretty :: Decl -> Doc ann

prettyList :: [Decl] -> Doc ann

Pretty Expr
Pretty instance for Expr - renders to valid Haskell source code.

Instance details

Methods

pretty :: Expr -> Doc ann

prettyList :: [Expr] -> Doc ann

Pretty Module
Pretty instance for Module - renders to valid Haskell source code.

Instance details

Methods

pretty :: Module -> Doc ann

prettyList :: [Module] -> Doc ann

Pretty Name

Instance details

Methods

pretty :: Name -> Doc ann

prettyList :: [Name] -> Doc ann

Pretty Pattern
Pretty instance for Pattern - renders to valid Haskell source code.

Instance details

Methods

pretty :: Pattern -> Doc ann

prettyList :: [Pattern] -> Doc ann

Pretty Type
Pretty instance for Type - renders to valid Haskell source code.

Instance details

Methods

pretty :: Type -> Doc ann

prettyList :: [Type] -> Doc ann

Pretty UnqualifiedName

Instance details

Methods

pretty :: UnqualifiedName -> Doc ann

prettyList :: [UnqualifiedName] -> Doc ann